+++ CHANGELOG.md
@@ -9,9 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added popup design and activation switch.
+- Added WinDivert initial setup.
 
 ### Fixed
 - Fixed downloads stopped because of app closure appear like they are still downloading when loaded.
+
+### Changed
+- Made WinDivert loop run in another thread to avoid conflicts.
 
 ## [v1.5.0] 03-09-2026